# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: XWiki API
  version: 0.8.0
  description: |
    面向人类与 AI Agent 的 Git-backed 文档管理系统。
    写入端点支持 Idempotency-Key 幂等。
servers:
  - url: /api/v1
security:
  - sessionCookie: []
  - bearerAuth: []
paths:
...
  /projects/{id}/locks:
    get:
      summary: 锁状态
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          $ref: "#/components/responses/LockResult"
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      summary: 获取页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          $ref: "#/components/responses/LockResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      summary: 释放页面锁
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/LockPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [released]
                properties:
                  released:
                    type: boolean
        "400":
          $ref: "#/components/responses/BadRequest"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: xwiki_session
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    ProjectID:
      name: id
      in: path
      required: true
      schema: { type: string }
...
    LockPath:
      name: path
      in: query
      required: true
      schema: { type: string }
      description: 要锁定的文档路径
  responses:
...
    BadRequest:
      description: 参数或校验错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: 冲突（revision_conflict / page_locked / idempotency_conflict / 名称冲突）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    LockResult:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              lock:
                $ref: "#/components/schemas/Lock"
                nullable: true
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: 机器可读错误码
            message:
              type: string
            request_id:
              type: string
...
    Lock:
      type: object
      required: [path, user_id, username, acquired_at, expires_at]
      properties:
        path:
          type: string
        user_id:
          type: string
        username:
          type: string
        acquired_at:
          type: string
        expires_at:
          type: string
...